spark_lottery/requirements.txt
pyyaml
sse
redis
web3
tornado
py-solc
schema
sqlalchemy
